Refactor project models structure
