+++ user_svc/alembic/env.py
@@ -6,7 +6,7 @@ from sqlalchemy.engine import Connection
 from sqlalchemy.ext.asyncio import async_engine_from_config
 
 from alembic import context
-from db.models import Base
+from models.db_models import Base
 from utils.utils import AppConfig
 
 app_config = AppConfig()
